refactor(P2.6 slice 2): move replace_with_clause_with_cte_reference_v2 into with_to_cte
Second sub-slice of P2.6 (REFACTORING_SAFETY_PLAN.md §5.1). Moves the ~1080-line
giant WITH→CTE-reference rewriter `replace_with_clause_with_cte_reference_v2`
(with its attached doc comment, old lines 11755–12936) verbatim from
plan_builder_utils.rs into render_plan/with_to_cte/mod.rs.

Byte-identical: the moved block diffs EXACTLY clean vs the original (no widening
needed — it was already `pub(crate)`). Re-exported `pub(crate) use
super::with_to_cte::replace_with_clause_with_cte_reference_v2;` so
`build_chained_with_match_cte_plan`'s single call site keeps resolving
unqualified. The fn's cross-module deps resolve via targeted imports added to
the module: `find_all_with_clauses_grouped` (from plan_builder_utils during the
transition; moves here in slice 3) and `plan_contains_with_clause` (from
plan_predicates, P2.4). A local `RenderPlanBuilderResult` alias mirrors the
private one in plan_builder_utils (same pattern as the P2.5 sibling modules
vlp_rewrite / cte_graph_joins_rewrite). Relocated the slice-1
`collect_property_accesses_tests` module to the end of mod.rs so production fns
precede it (clippy `items_after_test_module`).

Ratchet: the moved fn carries pre-existing schema-axis raw-flag reads
(`is_denormalized` ×4, `from/to_node_properties`, `from/to_label_column`,
`type_column`) — a documented pre-existing pattern this PR does not add to. The
baseline is regenerated as a **pure net-zero relocation**: every token count
that drops on plan_builder_utils.rs reappears on with_to_cte/mod.rs, totals
conserved (verified in the baseline diff). No new axis branching (§2.1).

Gates: 1612 lib + 529 integration (incl corpus_sweep + sql_golden) + 1 ratchet
+ 7 unit, 0 failures; goldens + corpus byte-identical (no churn); fmt/clippy
clean; warning-free --tests build. Next: P2.6 slice 3 (WITH-discovery cluster).
