+++ tests/rust/ratchet/baseline.txt
@@ -23,7 +23,7 @@ schema	from_label_column	src/render_plan/cte_extraction.rs	3
 schema	from_label_column	src/render_plan/cte_manager/mod.rs	6
 schema	from_label_column	src/render_plan/pattern_comprehension_sql.rs	1
 schema	from_label_column	src/render_plan/plan_builder_helpers.rs	9
-schema	from_label_column	src/render_plan/plan_builder_utils.rs	1
+schema	from_label_column	src/render_plan/with_to_cte/mod.rs	1
 schema	from_label_column	src/server/bolt_protocol/result_transformer.rs	3
 schema	from_label_column	src/sql_generator/emitters/clickhouse/multi_type_vlp_joins.rs	2
 schema	from_label_column	src/sql_generator/emitters/clickhouse/variable_length_cte.rs	12
@@ -50,9 +50,10 @@ schema	from_node_properties	src/render_plan/from_builder.rs	5
 schema	from_node_properties	src/render_plan/group_by_builder.rs	1
 schema	from_node_properties	src/render_plan/plan_builder.rs	3
 schema	from_node_properties	src/render_plan/plan_builder_helpers.rs	7
-schema	from_node_properties	src/render_plan/plan_builder_utils.rs	6
+schema	from_node_properties	src/render_plan/plan_builder_utils.rs	5
 schema	from_node_properties	src/render_plan/properties_builder.rs	16
 schema	from_node_properties	src/render_plan/select_builder.rs	2
+schema	from_node_properties	src/render_plan/with_to_cte/mod.rs	1
 schema	from_node_properties	src/server/bolt_protocol/result_transformer.rs	3
 schema	from_node_properties	src/sql_generator/emitters/clickhouse/json_builder.rs	1
 schema	from_node_properties	src/sql_generator/emitters/clickhouse/multi_type_vlp_joins.rs	4
@@ -94,11 +95,12 @@ schema	is_denormalized	src/render_plan/from_builder.rs	4
 schema	is_denormalized	src/render_plan/join_builder.rs	10
 schema	is_denormalized	src/render_plan/plan_builder.rs	10
 schema	is_denormalized	src/render_plan/plan_builder_helpers.rs	9
-schema	is_denormalized	src/render_plan/plan_builder_utils.rs	5
+schema	is_denormalized	src/render_plan/plan_builder_utils.rs	1
 schema	is_denormalized	src/render_plan/properties_builder.rs	3
 schema	is_denormalized	src/render_plan/render_expr.rs	1
 schema	is_denormalized	src/render_plan/select_builder.rs	2
 schema	is_denormalized	src/render_plan/vlp_rewrite.rs	4
+schema	is_denormalized	src/render_plan/with_to_cte/mod.rs	4
 schema	is_denormalized	src/sql_generator/emitters/clickhouse/to_sql_query.rs	1
 schema	is_denormalized	src/sql_generator/emitters/clickhouse/variable_length_cte.rs	41
 schema	is_fk_edge	src/procedures/apoc_meta_schema.rs	1
@@ -134,7 +136,7 @@ schema	to_label_column	src/render_plan/cte_extraction.rs	3
 schema	to_label_column	src/render_plan/cte_manager/mod.rs	8
 schema	to_label_column	src/render_plan/pattern_comprehension_sql.rs	1
 schema	to_label_column	src/render_plan/plan_builder_helpers.rs	9
-schema	to_label_column	src/render_plan/plan_builder_utils.rs	1
+schema	to_label_column	src/render_plan/with_to_cte/mod.rs	1
 schema	to_label_column	src/server/bolt_protocol/result_transformer.rs	3
 schema	to_label_column	src/sql_generator/emitters/clickhouse/multi_type_vlp_joins.rs	2
 schema	to_label_column	src/sql_generator/emitters/clickhouse/variable_length_cte.rs	17
@@ -161,9 +163,10 @@ schema	to_node_properties	src/render_plan/from_builder.rs	5
 schema	to_node_properties	src/render_plan/group_by_builder.rs	1
 schema	to_node_properties	src/render_plan/plan_builder.rs	1
 schema	to_node_properties	src/render_plan/plan_builder_helpers.rs	7
-schema	to_node_properties	src/render_plan/plan_builder_utils.rs	2
+schema	to_node_properties	src/render_plan/plan_builder_utils.rs	1
 schema	to_node_properties	src/render_plan/properties_builder.rs	11
 schema	to_node_properties	src/render_plan/select_builder.rs	3
+schema	to_node_properties	src/render_plan/with_to_cte/mod.rs	1
 schema	to_node_properties	src/server/bolt_protocol/result_transformer.rs	3
 schema	to_node_properties	src/sql_generator/emitters/clickhouse/json_builder.rs	1
 schema	to_node_properties	src/sql_generator/emitters/clickhouse/multi_type_vlp_joins.rs	4
@@ -186,7 +189,7 @@ schema	type_column	src/render_plan/cte_manager/mod.rs	6
 schema	type_column	src/render_plan/join_builder.rs	1
 schema	type_column	src/render_plan/pattern_comprehension_sql.rs	2
 schema	type_column	src/render_plan/plan_builder_helpers.rs	12
-schema	type_column	src/render_plan/plan_builder_utils.rs	1
+schema	type_column	src/render_plan/with_to_cte/mod.rs	1
 schema	type_column	src/server/bolt_protocol/result_transformer.rs	3
 schema	type_column	src/sql_generator/emitters/clickhouse/multi_type_vlp_joins.rs	2
 schema	type_column	src/sql_generator/emitters/clickhouse/variable_length_cte.rs	13
